% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{min_zero}
\alias{min_zero}
\title{Min Zero Function}
\usage{
min_zero(arg)
}
\arguments{
\item{arg}{A number}
}
\value{
\code{arg} if positive, otherwise 0.
}
\description{
This function just replaces negative values with zero.
}
\examples{
min_zero(27)
min_zero(-27)
}
