% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_bin.R
\name{pack_bin}
\alias{pack_bin}
\title{Pack Bin Function}
\usage{
pack_bin(item_dims, bin_dims, init_qty = 0, itr = 0)
}
\arguments{
\item{item_dims}{Vector containing item dimensions}

\item{bin_dims}{Vector containing bin dimensions}

\item{init_qty}{Number to track how many items fit in bin. Defaults to 0.}

\item{itr}{Number to tracks how many recursions have been run. Defaults to 0.}
}
\value{
Returns two values: how many items fit & how many iterations it took \code{pack_bin} to calculate.
}
\description{
Calculate the number of items that fit in a bin.
}
\examples{
pack_bin(c(7.5,5,.25),c(24,12,10))
pack_bin(c(14,12.5,3),c(24,12,10))
}
