% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{test_fit}
\alias{test_fit}
\title{Test Fit Function}
\usage{
test_fit(it_vec, to_vec)
}
\arguments{
\item{it_vec}{Vector containing item dimensions}

\item{to_vec}{Vector containing bin dimensions}
}
\description{
Test to see if an item fits in a bin, given the dimensions of each.
}
\examples{
test_fit(c(7.5,5,.25),c(24,12,10))
test_fit(c(14,12.5,3),c(24,12,10))
}
